Tree insert no longer commutable.
Revealed a bug, it started getting NullRefExceptions. Read something it should not have, perhaps. But also, since it conflicted a lot (and due to the bug, sometimes infinitely!) it kept the main read stamp open, slowing things down due to version leak. This must be improved, until then commutables should be quick.