One more for Thread.Abort.
Another finally block, to protect, at transaction opening, the moment of
inserting our version into the VersionList - it must be stored into the
_currentTransactionStartStamp too, otherwise it will never be removed
from the list, causing all copies from then on to be kept.

Seems silly to insert these finally blocks, but if it works, great. Still need
to write a test.