Added SERVER option.
If the compilation symbol SERVER is defined for Shielded.dll, it will use Monitor.Wait
and PulseAll for locking individual Shieldeds while writing in them, instead of SpinWait.
Works much better when you have a large number of competing threads, otherwise
slows things down slightly. And, surprisingly, the dictionary is faster that way even
with a low thread count.