Commutes simplified and further improved.
Basically, if one is compromised, they all must execute. If a main transaction is
"fixing", pin-pointing one of the results of some previous thing it did, then all results
of that previous thing must become pin-pointed, in order to be consistent with the
one result that the main transaction has read. Since the commutes can read whatever,
and we have no knowledge beforehand of this, then we must execute them all.

Apart from being very safe, this principle is also much simpler to implement.