fix - Removed the TransItems bag.
New test, TreePoolTest(), which runs the tree with a controlled number of
threads, revealed that after running for a while, Shield becomes slow. Last
part, running numItems empty transactions, would be reduced in speed to
merely 10k/s, but would be very fast if it was the first thing to execute in the test!
Removing the bag makes it run equally fast before or after the tree transactions.