A new test, to see if Shielded incurs too much overhead.
It would be very bad if repeatedly accessing an already accesed IShielded
would degrade performance. I presume that a more complex project, the kind
which stands most to gain from Shielded, would probably involve transactions
that will access the same IShielded over and over again. Composability is
a key advantage of STM, and the freedom to access whatever, wherever is
essential for it.

So, these new tests will repeatedly read from and write into the same
IShielded (specifically, an auto-generated subclass of a POCO) and measure
the time it takes to perform N of these transactions. Single access
transactions are also measured for comparison.

The results are very good!
- Reading and writing speeds are closely equal. I must admit, this was
unexpected.
- 200 read or write operations performed in one transaction make that
transaction only ~5% slower than a transaction which performs only one access.

All in all, there should be no fear of the fact that every field access goes through
the STM bookkeeping system. The first access incurs an expense, but the rest are
practically negligible.