Further improvements to the simple ops test.
Some InTransaction calls were in fact calling the overload which
returns a result. This overload creates a closure for that purpose, and
is much slower. The cost of an allocation (perhaps just on Mono) is
noticeable. One measurement of this is now included.

Configured like this, with 100000 transactions and 100 additional
operations, the test gives results which can be used to relatively easily
calculate the average contribution of each op to the running time.