Small optimization to commit process.
Items will not get ToList-ed, avoiding 1-2 list constructions per commit. The code in
DoCommit is now simpler. This may slightly increase time under main lock in
conflict situations, but the difference is not noticeable in any of the conflicting
tests. On the other hand, in SimpleOps this produces a noticeable, albeit small,
speed-up.