Fix commute double-committing.
Previous changeset created a minor issue. Items enlisted by commutes get
merged into the _localItems.Enlisted at the end of CommitCheck, regardless
of the outcome, and since the enumerable of enlisted items was no longer a
ToList-copy, it would after that contain the commutes as well. This would
lead to them being committed twice. (That has no negative consequences,
except for a slowdown.)