Fix for NullRefExc in ProxyGen.
A property with protected setter caused the ProxyGen to throw. I tried
adding support for a property with a public getter but protected setter,
however this is impossible with CodeDom - you cannot, for a property
declaration, separately specify the access modifiers for the getter and
the setter. Pretty lame. So, such a property will not be transactional.