IsolatedRun carries outer commutes inside.
Inside an IsolatedRun call we may enlist items for which the outer transaction
has defined commutes. Not only would the isolated transaction then see
invalid values of those fields, but it would also later allow the main transaction
to do the same, because later enlists would no longer be the first enlist of
that field, and would not trigger commute degeneration.