Added a simple prioritization scheme test.
The new test in PreCommitTests uses the PreCommit mechanism to
implement the simplest prioritization scheme. It will, when a thread
declares itself owner of the contested field, prevent all threads with
a different ManagedThreadId from committing into the field, instead
forcing them to retry until the owner releases the field. Test confirms
that without this a really slow transaction gets repeated due to faster
transactions conflicting with it, but with the protection active it
completes sucessfully in a single iteration. Also cool is that it was not
necessary to add any special code to the faster, conflicting transactions.
Merely by accessing the field they cause the PreCommit, that is
protecting it, to execute in their transaction. This means that most
transactions touching the field can be completely agnostic of any protection
involved and will none the less certainly respect it.