Removed the Rollback(false).
It was a dangerous function, allowing any method you make a call to, to quietly
kill your transaction, without your code having any idea that this happened.
You might think that the values you obained from your transaction run are ok,
since you are normally executing in code which is after the InTransaction call,
while in reality the transaction did not complete at all, and the values you
extracted are potentially invalid.

Shield.Rollback now receives no arguments, and will cause a repetition of the
outermost InTransaction call. If you want to stop a transaction, throw and catch
an exception yourself. This eliminates the problems with silent failure, as there
is no silent failure any more.