A little Shield.SideEffect safety.
The method lacked a check for transaction status and would throw a NullRef
exception if called out of transaction. It will now, if called out of transaction,
directly execute the onCommit action. This allows the possibilty of designing
a method which may consist of multiple transactions when executed alone, but
safely merges into any larger transaction when it gets called from it.

I believe the best way to design methods, though, is to either open your own
transaction covering the whole method, or to assert at beginning that you
already are in one. When they need a specific transaction status, methods in
Shielded will open their own if their action is equivalent to a read-only op, or
assert that they are already in one if they will produce any change.