Lock-free version tracking.
The VersionList is significantly changed to no longer be just a dummy list
of open readers. It now contains a list of all written versions, between the
oldest one being read and the newest one written in. It is now in charge of
trimming, and it can do it without any additional data structures.

The only remaining lock statements (if not compiled with SERVER flag) are now
the ones used to safely remove items from concurrent dictionaries.

Also, perhaps surprisingly, this version is actually faster than the simple
lock-using one.