Fix #4 - NullRefExc due to too old reader ticket
* The method GetReaderTicket in class VersionList performs two reads
of the _current static field in the same class. Since there is an
Interlocked operation between them, this is safe with respect to the
hardware. Interlocked ops emit fences. However, it is not enough to
stop the JIT compiler from "optimizing" those two reads into just
one. Marking the field volatile stops such optimizations.

Many thanks to @ocozalp for testing the fixed version, since I was
unable to reproduce the issue on my machine.

* Also included is a small improvement to the DictionaryTest in the
console tests project.