Another fix for #4, and some optimizations.
* Closed another extremely unlikely but possible race condition
between trimming and issuing reader tickets. Also improved the way
trimming and issuing reader tickets interact. They now conflict
directly, over one field, making it impossible for these errors to
happen, which can now easily be concluded from the code.
* Trimming was needlessly creating hash sets, and redoing trimming
which was already done. This was not an error, just a waste of time.
* SimpleHashSet now supports trimming better by recognizing when
it is being merged with a List and doing a for loop instead of
foreach.