Fix for the dictionary expanding problem
Since the dictionary presents itself as just one field to the Shield
class, the Shield class could properly check for safety when someone
is accesing it inside a WhenCommitting subscription. It was possible,
if the dict was written, to perform additional writes under new keys.
That was unsafe, since those keys were not checked and locked.
This changeset closes this issue and adds an exhaustive confirmation
test.