Fix for issue with Conditionals
The Remover method was as of recently overzealously removing "this"
from the lists of subscriptions, when it should only remove one
entry, to match the one addition that its thread made before. This
issue would affect volatile conditionals, which change their access
pattern between calls to the test lambda, and could result in the
conditional being effectively deactivated.