Bugfix for old version trimming
A subtle one. It was possible for us to leave the loop in
VersionList.TrimCopies due to Changes being null, but then after the
loop we set the Changes to Enumerable.Empty. This is unsafe, and may
result in forgotten fields. Not critical, since they would eventually
get trimmed, but might result in needlessly prolonged holding of old
versions.