Corrected the classification to "mostly lock-free"
According to academic definitions, obstruction-freedom would mean that two (or more) transactions might forever loop, constantly making each other roll back. Shielded guarantees that conflicts can only occur when at least one of them commits, thus guaranteeing system-wide progress. Of course, this is not fully correct due to the commit check being done under a lock. The LockFree branch gets close, but even it does not satisfy it - a thread that stops during the check, or between it and actual writing, keeps fields (partially) locked. A mechanism for threads to "help" each other finish the check and write would be needed to fully satisfy the definition. But I believe even the master branch, as it is, gains much from getting close to the ideal.