Version 4.0.0; async commit
Method Shield.RunToCommit runs a transaction until it passes the
commit check. It returns a continuation object which can be later
used, from any thread, to commit or rollback the transaction.
The fields remain locked until the continuation is used. This
required a refactoring, to enable using the transaction info from
a different thread.

Breaking change: LocalStorage is no more. It's now called
TransactionalStorage, and works only inside transactions. Storage
is now automatically released when the transaction completes.