Safer WhenCommitting
The subscriptions can no longer cause a rollback, which was a safety
issue before as well. They will now all get a chance execute, and an
AggregateException will maybe be thrown after all of them have
executed. The AggregateException bubbles out of either a
Shield.InTransaction or CommitContinuation.Commit call.