TransactionalStorage now internal
Since it stopped being just a thread-local storage, and became bound
to the current transaction context, it has had a serious pitfall - if
your data landed in the private fields, it would not get automatically
cleaned up at transaction closing! It is now internal, and slightly
optimized since all it's use cases are now fully known.