Fix for ShieldedLocal
It does not use the internal TransactionalStorage anymore. The
internal class has private fields for faster use by one thread, but
because of that it depends on IShielded implementors calling
Release on it. ShieldedLocal does not call Release on normal
transaction close, so the private fields would forever keep their
first value. ShieldedLocal now uses only the dictionary from the
current context, which makes it safe.