Fix for CommitContinuation safety
Since the Storage is an ordinary dictionary, it was unsafe to allow
InContext to run in parallel with any other thread trying a commit or
rollback, or even just using InContext, particularly if changing any
values is allowed. Now all three use a simple lock to synchronize,
which also means that (Try)Commit and (Try)Rollback calls may block
if another thread is stuck in InContext, or in some WhenCommitting or
SyncSideEffect code. This seems like the only safe way.