Fix for possible leak when TransExc gets caught
TransactionLoop would, when seeing _rollback flag set, go directly to
DoCheckFailed, but this one assumes that the fields are already
rolled back, as the check normally guarantees. This would leave the
private fields in TransactionalStorage instances occupied, causing a
small memory leak (may happen max. once per field), and of course the
loss of the fast local fields.