Fix for ProxyGen reference resolving
When creating the CodeDOM for the proxy class, the old
version would only check the base type and interfaces to
find any assemblies to add to the CodeDOM compilation
unit. However, fields may also use types which are defined
in another assembly!

To fix this, I had to write a class which resolves
dependencies... I tried using all loaded assemblies, but the
dependencies are not loaded yet if you don't explicitly use
the type you're trying to make a proxy for. E.g. the test here
that runs through all types in an assembly would fail in that
case, because one type's field uses a type from another, as
yet not loaded assembly. Even doing
Activator.CreateInstance would not force .NET to load the
dependencies of that type.