Porting fixes and improvements from Shielded.Gossip
The other project has a version of Shielded targetting
.NET Standard 2. In the process of developing there, I
made some improvements to the Shielded library which
I'm porting over here now.
* TransException will now be caught by the Shield class
even if it is wrapped in a TargetInvocationException or
AggregateException. We search those recursively. Please
do not take this support for AggregateException as a sign
that it's OK to use Tasks inside transactions - it is not. It's
just that Shielded.Gossip uses Tasks for distributed
transactions, and when nesting those, Task class insists on
wrapping any exceptions into an AggrExc, even if it is just
one Task, which did not do anything async...
* Shield.IsTransException allows easy checking if an
Exception is a TransException, incl. the above logic of
recursively looking inside TargetInvocationExceptions
and AggregateExceptions.
* Shield.ReadOldState<T> added, which simply returns
whatever the lambda returned, for easier use.
* Fix for a bug - if a PreCommit adds a new commute,
such a commute would not have been executed if there
were no commutes subscribed in the main transaction.
* Minor bug - ShieldedDict(Nc).Clear did not check if
we were in a transaction, and would work, but only if
the dictionary is empty.
* Added ShieldedLocal.GetValueOrDefault.
* Added ShieldedTree(Nc).Remove overload which
accepts a key and value. Old version had this functionality,
but only receiving a KeyValuePair. This is nicer to use.
* Added ShieldedTree(Nc).IsEmpty, which is much faster
than any enumerator when the tree is not empty.

This is also the first time I'm using new C# features in this
project. Currently only the exception filter, for checking if
we should catch an exception.