Port fixes from Shielded.Gossip
* SyncSideEffect which reads from commuted fields
might cause the fields to throw a TransException, because
the main transaction read version is lower than what was
last written in them. But being commutes means that they
delt with this and the commutable op did in fact read
that newer version. Thus, the rollback caused by the throw
was wrong. Fixed.
* ShieldedLocal.GetValueOrDefault receives an argument
for the value to return if the local is empty.