Fix minor ShieldedDict clean-up issues
* We will no longer add a new Empty item if there was nothing
in the underlying dictionary before, since it's useless.
* Fix in clean-up, "<=" instead of "<", since the version in each
copy list entry is not the version of the old copies, it's the
version that obsoleted them! The copies are thus safe to remove
since their version is definitely smaller than that.
* The Remove method had a small bug. If the same transaction
would set and then remove a key, the item that would end up in
the dictionary would have Empty == true and Value != null.